+++ R/lfo.R
@@ -54,11 +54,8 @@ lfo <- function(x, ...) {
 #'   * `k_thres`: The Pareto k threshold used.
 #' @examples
 #' \donttest{
-#' lfo_approx <- lfo(fit_single_treated, L = 10)
+#' lfo_approx <- lfo(fit_single_treated, L = 20)
 #' lfo_approx
-#'
-#' lfo_exact <- lfo(fit_single_treated, L = 10, exact = TRUE)
-#' lfo_exact
 #' }
 lfo.bscmfit <- function(
   x,